undo refactor
need to undo what i had refactored
